+++ core/base-service/base-non-memory-caching.js
@@ -4,6 +4,7 @@ const makeBadge = require('../../gh-badges/lib/make-badge')
 const BaseService = require('./base')
 const { setCacheHeaders } = require('./cache-headers')
 const { makeSend } = require('./legacy-result-sender')
+const coalesceBadge = require('./coalesce-badge')
 
 // Badges are subject to two independent types of caching: in-memory and
 // downstream.
@@ -35,8 +36,12 @@ module.exports = class NonMemoryCachingBaseService extends BaseService {
         queryParams
       )
 
-      const badgeData = this._makeBadgeData(queryParams, serviceData)
-      // The final capture group is the extension.
+      const badgeData = coalesceBadge(
+        queryParams,
+        serviceData,
+        this.defaultBadgeData,
+        this
+      ) // The final capture group is the extension.
       const format = match.slice(-1)[0]
       badgeData.format = format
