+++ core/base-service/base-static.js
@@ -8,6 +8,7 @@ const {
   setCacheHeadersForStaticResource,
 } = require('./cache-headers')
 const { makeSend } = require('./legacy-result-sender')
+const coalesceBadge = require('./coalesce-badge')
 
 module.exports = class BaseStaticService extends BaseService {
   static register({ camp }, serviceConfig) {
@@ -33,8 +34,12 @@ module.exports = class BaseStaticService extends BaseService {
         queryParams
       )
 
-      const badgeData = this._makeBadgeData(queryParams, serviceData)
-      // The final capture group is the extension.
+      const badgeData = coalesceBadge(
+        queryParams,
+        serviceData,
+        this.defaultBadgeData,
+        this
+      ) // The final capture group is the extension.
       const format = match.slice(-1)[0]
       badgeData.format = format
