+++ core/base-service/base.spec.js
@@ -4,7 +4,6 @@ const Joi = require('joi')
 const { expect } = require('chai')
 const { test, given, forCases } = require('sazerac')
 const sinon = require('sinon')
-const { getShieldsIcon } = require('../../lib/logos')
 const trace = require('./trace')
 
 const {
@@ -367,245 +366,6 @@ describe('BaseService', function() {
           color: 'red',
           message: 'invalid parameter',
         })
-      })
-    })
-  })
-
-  describe('_makeBadgeData', function() {
-    describe('Overrides', function() {
-      it('overrides the label', function() {
-        const badgeData = DummyService._makeBadgeData(
-          { label: 'purr count' },
-          { label: 'purrs' }
-        )
-        expect(badgeData.text).to.deep.equal(['purr count', 'n/a'])
-      })
-
-      it('overrides the label color', function() {
-        const badgeData = DummyService._makeBadgeData(
-          { colorA: '42f483' },
-          { color: 'green' }
-        )
-        expect(badgeData.labelColor).to.equal('42f483')
-      })
-
-      it('overrides the color', function() {
-        const badgeData = DummyService._makeBadgeData(
-          { colorB: '10ADED' },
-          { color: 'red' }
-        )
-        expect(badgeData.color).to.equal('10ADED')
-      })
-
-      it('converts a query-string numeric color to a string', function() {
-        const badgeData = DummyService._makeBadgeData(
-          // Scoutcamp converts numeric query params to numbers.
-          { colorB: 123 },
-          { color: 'green' }
-        )
-        expect(badgeData.color).to.equal('123')
-      })
-
-      it('does not override the color in case of an error', function() {
-        const badgeData = DummyService._makeBadgeData(
-          { colorB: '10ADED' },
-          { isError: true, color: 'lightgray' }
-        )
-        expect(badgeData.color).to.equal('lightgray')
-      })
-
-      it('overrides the logo', function() {
-        const badgeData = DummyService._makeBadgeData(
-          { logo: 'github' },
-          { namedLogo: 'appveyor' }
-        )
-        // .not.be.empty for confidence that nothing has changed with `getShieldsIcon()`.
-        expect(badgeData.logo).to.equal(getShieldsIcon({ name: 'github' })).and
-          .not.be.empty
-      })
-
-      it('overrides the logo with a color', function() {
-        const badgeData = DummyService._makeBadgeData(
-          { logo: 'github', logoColor: 'blue' },
-          { namedLogo: 'appveyor' }
-        )
-        expect(badgeData.logo).to.equal(
-          getShieldsIcon({ name: 'github', color: 'blue' })
-        ).and.not.be.empty
-      })
-
-      it("when the logo is overridden, it ignores the service's logo color, position, and width", function() {
-        const badgeData = DummyService._makeBadgeData(
-          { logo: 'github' },
-          {
-            namedLogo: 'appveyor',
-            logoColor: 'red',
-            logoPosition: -3,
-            logoWidth: 100,
-          }
-        )
-        expect(badgeData.logo).to.equal(getShieldsIcon({ name: 'github' })).and
-          .not.be.empty
-      })
-
-      it("overrides the service logo's color", function() {
-        const badgeData = DummyService._makeBadgeData(
-          { logoColor: 'blue' },
-          { namedLogo: 'github', logoColor: 'red' }
-        )
-        expect(badgeData.logo).to.equal(
-          getShieldsIcon({ name: 'github', color: 'blue' })
-        ).and.not.be.empty
-      })
-
-      it('overrides the logo with custom svg', function() {
-        const logoSvg = 'data:image/svg+xml;base64,PHN2ZyB4bWxu'
-        const badgeData = DummyService._makeBadgeData(
-          { logo: logoSvg },
-          { namedLogo: 'appveyor' }
-        )
-        expect(badgeData.logo).to.equal(logoSvg)
-      })
-
-      it('ignores the color when custom svg is provided', function() {
-        const logoSvg = 'data:image/svg+xml;base64,PHN2ZyB4bWxu'
-        const badgeData = DummyService._makeBadgeData(
-          { logo: logoSvg, logoColor: 'brightgreen' },
-          { namedLogo: 'appveyor' }
-        )
-        expect(badgeData.logo).to.equal(logoSvg)
-      })
-
-      it('overrides the logoWidth', function() {
-        const badgeData = DummyService._makeBadgeData({ logoWidth: 20 }, {})
-        expect(badgeData.logoWidth).to.equal(20)
-      })
-
-      it('overrides the logoPosition', function() {
-        const badgeData = DummyService._makeBadgeData({ logoPosition: -10 }, {})
-        expect(badgeData.logoPosition).to.equal(-10)
-      })
-
-      it('overrides the links', function() {
-        const badgeData = DummyService._makeBadgeData(
-          { link: 'https://circleci.com/gh/badges/daily-tests' },
-          {
-            link:
-              'https://circleci.com/workflow-run/184ef3de-4836-4805-a2e4-0ceba099f92d',
-          }
-        )
-        expect(badgeData.links).to.deep.equal([
-          'https://circleci.com/gh/badges/daily-tests',
-        ])
-      })
-
-      it('overrides the template', function() {
-        const badgeData = DummyService._makeBadgeData({ style: 'pill' }, {})
-        expect(badgeData.template).to.equal('pill')
-      })
-
-      it('overrides the cache length', function() {
-        const badgeData = DummyService._makeBadgeData(
-          { style: 'pill' },
-          { cacheSeconds: 123 }
-        )
-        expect(badgeData.cacheLengthSeconds).to.equal(123)
-      })
-    })
-
-    describe('Service data', function() {
-      it('applies the service message', function() {
-        const badgeData = DummyService._makeBadgeData({}, { message: '10k' })
-        expect(badgeData.text).to.deep.equal(['cat', '10k'])
-      })
-
-      it('preserves an empty label', function() {
-        const badgeData = DummyService._makeBadgeData(
-          {},
-          { label: '', message: '10k' }
-        )
-        expect(badgeData.text).to.deep.equal(['', '10k'])
-      })
-
-      it('applies a numeric service message', function() {
-        // While a number of badges use this, in the long run we may want
-        // `render()` to always return a string.
-        const badgeData = DummyService._makeBadgeData({}, { message: 10 })
-        expect(badgeData.text).to.deep.equal(['cat', 10])
-      })
-
-      it('applies the service color', function() {
-        const badgeData = DummyService._makeBadgeData({}, { color: 'red' })
-        expect(badgeData.color).to.equal('red')
-      })
-
-      it('applies the named logo', function() {
-        const badgeData = DummyService._makeBadgeData(
-          {},
-          { namedLogo: 'github' }
-        )
-        // .not.be.empty for confidence that nothing has changed with `getShieldsIcon()`.
-        expect(badgeData.logo).to.equal(getShieldsIcon({ name: 'github' })).and
-          .not.to.be.empty
-      })
-
-      it('applies the named logo with color', function() {
-        const badgeData = DummyService._makeBadgeData(
-          {},
-          { namedLogo: 'github', logoColor: 'blue' }
-        )
-        expect(badgeData.logo).to.equal(
-          getShieldsIcon({ name: 'github', color: 'blue' })
-        ).and.not.to.be.empty
-      })
-
-      it('applies the logo width', function() {
-        const badgeData = DummyService._makeBadgeData(
-          {},
-          { namedLogo: 'github', logoWidth: 275 }
-        )
-        expect(badgeData.logoWidth).to.equal(275)
-      })
-
-      it('applies the logo position', function() {
-        const badgeData = DummyService._makeBadgeData(
-          {},
-          { namedLogo: 'github', logoPosition: -10 }
-        )
-        expect(badgeData.logoPosition).to.equal(-10)
-      })
-
-      it('applies the service label color', function() {
-        const badgeData = DummyService._makeBadgeData({}, { labelColor: 'red' })
-        expect(badgeData.labelColor).to.equal('red')
-      })
-    })
-
-    describe('Defaults', function() {
-      it('uses the default label', function() {
-        const badgeData = DummyService._makeBadgeData({}, {})
-        expect(badgeData.text).to.deep.equal(['cat', 'n/a'])
-      })
-
-      it('uses the default color', function() {
-        const badgeData = DummyService._makeBadgeData({}, {})
-        expect(badgeData.color).to.equal('lightgrey')
-      })
-
-      it('provides no default label color', function() {
-        const badgeData = DummyService._makeBadgeData({}, {})
-        expect(badgeData.labelColor).to.be.undefined
-      })
-
-      it('when not a social badge, ignores the default named logo', function() {
-        const badgeData = DummyService._makeBadgeData({}, {})
-        expect(badgeData.logo).to.be.undefined
-      })
-
-      it('when a social badge, uses the default named logo', function() {
-        const badgeData = DummyService._makeBadgeData({ style: 'social' }, {})
-        expect(badgeData.logo).to.equal(getShieldsIcon({ name: 'appveyor' }))
-          .and.not.be.empty
       })
     })
   })
