+++ core/base-service/transform-example.js
@@ -2,6 +2,7 @@
 
 const Joi = require('joi')
 const pathToRegexp = require('path-to-regexp')
+const coalesceBadge = require('./coalesce-badge')
 
 const optionalObjectOfKeyValues = Joi.object().pattern(
   /./,
@@ -142,7 +143,12 @@ function transformExample(inExample, index, ServiceClass) {
     const {
       text: [label, message],
       color,
-    } = ServiceClass._makeBadgeData({}, staticPreview)
+    } = coalesceBadge(
+      {},
+      staticPreview,
+      ServiceClass.defaultBadgeData,
+      ServiceClass
+    )
     preview = { label, message: `${message}`, color }
   } else {
     preview = {
